# Low-hanging opt: avoid allocation on default sampler