A low-risk and simple change to avoid allocations from the typical sampler usage: just avoiding creating and calling it on the typical case. We can look at removing the samplers since we go for high-fidelity but that will require a major version bump since it will remove some current capability (granted that I'm not aware of usage for it).