@pjanotti 2021-06-07
I used the no tagging test for NLog 4.6 as a proxy for the effects of this change. Notice mean duration and allocated reductions.

### Original

|              Method |         Categories |      Mean |     Error |    StdDev | Ratio |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|-------------------- |------------------- |----------:|----------:|----------:|------:|-------:|-------:|-------:|----------:|
|           NoTagging | NoActiveSpanSwitch |  8.191 us | 0.3922 us | 1.1127 us |  1.00 | 0.5798 | 0.2136 | 0.0916 |   3.53 KB |
|                     |                    |           |           |           |       |        |        |        |           |
| NoTaggingSpanSwitch |   ActiveSpanSwitch | 11.084 us | 0.3181 us | 0.9328 us |  1.00 | 1.1444 | 0.4883 | 0.1221 |   4.78 KB |

### Optimized

|              Method |         Categories |      Mean |     Error |    StdDev | Ratio |  Gen 0 |  Gen 1 |  Gen 2 | Allocated |
|-------------------- |------------------- |----------:|----------:|----------:|------:|-------:|-------:|-------:|----------:|
|           NoTagging | NoActiveSpanSwitch |  6.444 us | 0.2690 us | 0.7846 us |  1.00 | 0.5798 | 0.2899 | 0.0305 |   3.41 KB |
|                     |                    |           |           |           |       |        |        |        |           |
| NoTaggingSpanSwitch |   ActiveSpanSwitch | 10.351 us | 0.3255 us | 0.9444 us |  1.00 | 0.9766 | 0.4883 | 0.1221 |   4.66 KB |